requirements.txt
flask
pyzbar
pillow
exifread
